+++ src/xulbux/__init__.py
@@ -7,11 +7,26 @@ __copyright__ = "Copyright (c) 2024 XulbuX"
 __url__ = "https://github.com/XulbuX/PythonLibraryXulbuX"
 __description__ = "A Python library which includes lots of helpful classes, types, and functions aiming to make common programming tasks simpler."
 
-__all__ = ["Code", "Color", "Console", "Data", "EnvPath", "File", "FormatCodes", "Json", "Path", "Regex", "String", "System"]
+__all__ = [
+    "Code",
+    "Color",
+    "Console",
+    "Data",
+    "EnvPath",
+    "File",
+    "FormatCodes",
+    "Json",
+    "Path",
+    "ProgressBar",
+    "Regex",
+    "String",
+    "System",
+]
 
 from .code import Code
 from .color import Color
 from .console import Console
+from .console import ProgressBar
 from .data import Data
 from .env_path import EnvPath
 from .file import File